+++ react/ListPriceRange.tsx
@@ -52,7 +52,7 @@ function ListPriceRange({
   noRangeMessage = messages.noRangeMessageDefault.id,
   markers = [],
 }: Props) {
-  const handles = useCssHandles(CSS_HANDLES)
+  const { handles } = useCssHandles(CSS_HANDLES)
   const productContextValue = useProduct()
 
   const priceRange = productContextValue?.product?.priceRange