+++ react/Savings.tsx
@@ -35,7 +35,7 @@ interface Props {
 }
 
 function Savings({ message = messages.default.id, markers = [] }: Props) {
-  const handles = useCssHandles(CSS_HANDLES)
+  const { handles } = useCssHandles(CSS_HANDLES)
   const productContextValue = useProduct()
   const productSummaryValue = ProductSummaryContext.useProductSummary()