+++ react/SpotPriceSavings.tsx
@@ -37,7 +37,7 @@ function SpotPriceSavings({
   message = messages.default.id,
   markers = [],
 }: Props) {
-  const handles = useCssHandles(CSS_HANDLES)
+  const { handles } = useCssHandles(CSS_HANDLES)
   const productContextValue = useProduct()
 
   const availableSeller = getFirstAvailableSeller(