+++ react/components/PriceLoadingSpinner/index.tsx
@@ -6,7 +6,7 @@ import styles from './styles.css'
 const CSS_HANDLES = ['priceLoading'] as const
 
 const PriceLoadingSpinner = () => {
-  const handles = useCssHandles(CSS_HANDLES)
+  const { handles } = useCssHandles(CSS_HANDLES)
 
   return (
     <div className={`${handles.priceLoading} flex items-end w-100 h1 pr6`}>